prj: pretest:e2e now generates prisma and migrates only for services with schema
This change is done to allow services without prisma to run e2e. C solution slows down execution a bit as it runs generate and migration for all services (and not only for the one that we are running the e2e tests for) but it's done this way to simplify the flow.